+++ lintrules/src/main/java/biz/davidpearson/android/customlintrules/PrintStackTraceDetector.java
@@ -1,5 +1,6 @@
 package biz.davidpearson.android.customlintrules;
 
+import com.android.tools.lint.client.api.JavaEvaluator;
 import com.android.tools.lint.detector.api.Category;
 import com.android.tools.lint.detector.api.Detector;
 import com.android.tools.lint.detector.api.Implementation;
@@ -7,12 +8,11 @@ import com.android.tools.lint.detector.api.Issue;
 import com.android.tools.lint.detector.api.JavaContext;
 import com.android.tools.lint.detector.api.Scope;
 import com.android.tools.lint.detector.api.Severity;
-import com.intellij.psi.JavaElementVisitor;
 import com.intellij.psi.PsiMethod;
-import com.intellij.psi.PsiMethodCallExpression;
-import com.intellij.psi.PsiReferenceExpression;
 
-import java.util.Arrays;
+import org.jetbrains.uast.UCallExpression;
+
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -20,7 +20,7 @@ import java.util.List;
  * <p>
  * a detector that looks for the use of <code>Throwable.printStackTrace()</code>
  */
-public final class PrintStackTraceDetector extends Detector implements Detector.JavaPsiScanner {
+public final class PrintStackTraceDetector extends Detector implements Detector.UastScanner {
 
     public static final Issue ISSUE_PRINT_STACK_TRACE =
             Issue.create("PrintStackTrace", "Use of printStackTrace",
@@ -30,21 +30,19 @@ public final class PrintStackTraceDetector extends Detector implements Detector.
 
     @Override
     public List<String> getApplicableMethodNames() {
-        return Arrays.asList("printStackTrace");
+        return Collections.singletonList("printStackTrace");
     }
 
     @Override
-    public void visitMethod(JavaContext context, JavaElementVisitor visitor,
-                            PsiMethodCallExpression call, PsiMethod method) {
+    public void visitMethodCall(JavaContext context, UCallExpression call, PsiMethod method) {
 
+        String methodName = call.getMethodName();
+        JavaEvaluator evaluator = context.getEvaluator();
 
-        PsiReferenceExpression methodExpression = call.getMethodExpression();
-
-        String fullyQualifiedMethodName = methodExpression.getQualifiedName();
-
-        if ("java.lang.Throwable.printStackTrace".equals(fullyQualifiedMethodName)) {
-            context.report(ISSUE_PRINT_STACK_TRACE, methodExpression,
-                    context.getLocation(methodExpression), "Do not use printStackTrace");
+        if ("printStackTrace".equals(methodName) &&
+                evaluator.isMemberInClass(method, "java.lang.Throwable")) {
+            context.report(ISSUE_PRINT_STACK_TRACE, call,
+                    context.getLocation(call), "Do not use printStackTrace");
             return;
         }
     }