+++ lintrules/src/test/java/biz/davidpearson/android/customlintrules/PrintStackTraceDetectorTest.java
@@ -4,25 +4,24 @@ import com.android.tools.lint.checks.infrastructure.LintDetectorTest;
 import com.android.tools.lint.detector.api.Detector;
 import com.android.tools.lint.detector.api.Issue;
 
-import org.intellij.lang.annotations.Language;
+import org.junit.Test;
 
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
-import static org.fest.assertions.api.Assertions.assertThat;
-
+/**
+ * requires ANDROID_HOME to be set in the environment
+ */
 public class PrintStackTraceDetectorTest extends LintDetectorTest {
 
+    @Test
     public void testPrintStackTracePresent() throws Exception {
 
-        System.out.println("\n\n\ntestPrintStackTracePresent called\n\n\n");
-
-        //@formatter:off
-        @Language("JAVA") String source = ""
+        lint().files(java(""
                 + "package foo;\n"
                 + "import android.content.Context;\n"
                 + "import android.os.Bundle;\n"
-                + "import android.support.v7.app.AppCompatActivity;\n"
+                + "import androidx.appcompat.app.AppCompatActivity;\n"
                 + "public class MainActivity extends AppCompatActivity {\n"
                 + "    @Override\n"
                 + "    protected void onCreate(Bundle savedInstanceState) { \n"
@@ -34,20 +33,11 @@ public class PrintStackTraceDetectorTest extends LintDetectorTest {
                 + "            e.printStackTrace();\n"
                 + "        }\n"
                 + "    }\n"
-                + "}";
-
-        @Language("JAVA") String foo = "";
-
-
-        assertThat(lintProject(java(source))).isEqualTo("src/foo/Example.java:13: "
-                + "Example.java:13: Do not use printStackTrace [PrintStackTrace]\n"
-                + "    } catch (Exception e) {\n"
-                + "        e.printStackTrace();\n"
-                + "          ~~~~~~~~~~~~~~~\n"
-                + "    }\n"
-                + "1 errors, 0 warnings\n");
-
-        //@formatter:on
+                + "}")).run()
+                .expect("src/foo/MainActivity.java:13: Error: Do not use printStackTrace [PrintStackTrace]\n"
+                        + "            e.printStackTrace();\n"
+                        + "            ~~~~~~~~~~~~~~~~~~~\n"
+                        + "1 errors, 0 warnings\n");
     }
 
     @Override
@@ -57,7 +47,7 @@ public class PrintStackTraceDetectorTest extends LintDetectorTest {
 
     @Override
     protected List<Issue> getIssues() {
-        return Arrays.asList(PrintStackTraceDetector.ISSUE_PRINT_STACK_TRACE);
+        return Collections.singletonList(PrintStackTraceDetector.ISSUE_PRINT_STACK_TRACE);
     }
 
     @Override