+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Change Log
 Version 0.1.0 *(2017-06-11)*
 ----------------------------
 Initial release.
+- build using Android Lint Tools 25.3.0